refactor: split backend peak files into device, compute kernel, and orchestrator
Split each backend's monolithic *_peak.cpp into focused files following a
consistent cross-backend pattern:

Vulkan (1235 -> 345 lines in vk_peak.cpp):
  - vulkan_device.cpp: VulkanDevice class, 4-step init() broken into helpers
  - compute_kernel.cpp: vkPeak::runComputeKernel() shared compute driver

CUDA (586 -> 250 lines in cuda_peak.cpp):
  - cuda_device.cpp: CudaDevice class + NVRTC compilation/caching
  - compute_kernel.cpp: CudaPeak::runComputeKernel()

Metal (540 -> 180 lines in mtl_peak.mm):
  - mtl_device.mm: MetalDevice class + copyClpeakMetalDevices()
  - mtl_utils.mm: mtlGetLibrary/mtlGetPipeline caching helpers
  - compute_kernel.mm: runComputeKernel + mtlRunDispatches

OpenCL (573 -> 350 lines in cl_peak.cpp):
  - cl_kernels.cpp: kernel string includes + accessor functions
  - compute_test.cpp: clPeak::runComputeTest()

All CMakeLists.txt and AGENTS.md updated accordingly.
