+++ src/metal/CMakeLists.txt
@@ -19,6 +19,9 @@ endif()
 
 add_library(peak_metal STATIC
     mtl_peak.mm
+    mtl_device.mm
+    mtl_utils.mm
+    compute_kernel.mm
     mtl_blas.mm
     compute_float.mm
     compute_int.mm
@@ -47,7 +50,7 @@ target_compile_definitions(peak_common PRIVATE ENABLE_METAL)
 
 # Compile .mm files with ObjC ARC -- saves us from sprinkling release
 # calls on autoreleased Metal objects.
-set_source_files_properties(mtl_peak.mm mtl_blas.mm compute_float.mm compute_int.mm simdgroup.mm global_bandwidth.mm local_bandwidth.mm image_bandwidth.mm kernel_latency.mm atomic_throughput.mm
+set_source_files_properties(mtl_peak.mm mtl_device.mm mtl_utils.mm compute_kernel.mm mtl_blas.mm compute_float.mm compute_int.mm simdgroup.mm global_bandwidth.mm local_bandwidth.mm image_bandwidth.mm kernel_latency.mm atomic_throughput.mm
     PROPERTIES COMPILE_FLAGS "-fobjc-arc")
 
 # Embed .metal kernel sources as C++ raw-string literals
