+++ src/vulkan/AGENTS.md
@@ -5,9 +5,11 @@ and GLSL compute shaders (in `shaders/`).  Built as `peak_vulkan` static library
 
 ## Quick Lookups
 
-- Looking for the main class? → `vk_peak.cpp`
-- Looking for instance/device init? → `vk_peak.cpp` (`initInstance`, `VulkanDevice::init`)
-- Looking for the unified compute kernel runner? → `vk_peak.cpp` (`runComputeKernel`)
+- Looking for the main class / orchestrator? → `vk_peak.cpp`
+- Looking for VulkanDevice class (logical device, buffers, pipelines)? → `vulkan_device.cpp`
+- Looking for instance init? → `vk_peak.cpp` (`initInstance`)
+- Looking for the unified compute kernel runner? → `compute_kernel.cpp` (`runComputeKernel`)
+- Looking for kernel timing/calibration? → `vk_peak.cpp` (`runKernel`)
 - Looking for GLSL shader sources? → `shaders/*.comp`
 - Looking for shader compilation logic? → `cmake/CompileShaders.cmake`
 - Looking for FP compute benchmarks? → `compute_float.cpp`
@@ -21,7 +23,9 @@ and GLSL compute shaders (in `shaders/`).  Built as `peak_vulkan` static library
 
 | File | Purpose |
 |------|---------|
-| `vk_peak.cpp` | `vkPeak` + `VulkanDevice` classes: ctor, `applyOptions()`, `runAll()`, `runComputeKernel()`, `run_kernel()`, inventory |
+| `vk_peak.cpp` | `vkPeak` class: ctor, `applyOptions()`, `initInstance()`, `cleanup()`, `runKernel()`, `runAll()`, `enumerate()`, `printInventory()` |
+| `vulkan_device.cpp` | `VulkanDevice` class: `init()` (4-step: basic info → CU count → optional features → logical device), `cleanup()`, `createBuffer()`, `createComputePipeline()`, `submitAndWait()`, `zeroBuffer()` |
+| `compute_kernel.cpp` | `vkPeak::runComputeKernel()` — shared compute-peak driver: buffer/descriptor/pipeline scaffolding used by all `runCompute*` wrappers |
 | `compute_float.cpp` | `runComputeSP`, `runComputeHP`, `runComputeDP`, `runComputeMP`, `runComputeBF16` |
 | `compute_int.cpp` | `runComputeInt32`, `runComputeInt8DP`, `runComputeInt4Packed` |
 | `coopmat.cpp` | `runCoopMatrix` — cooperative matrix (tensor-core) umbrella |
@@ -39,4 +43,5 @@ and GLSL compute shaders (in `shaders/`).  Built as `peak_vulkan` static library
 - If you add a new benchmark → add it to the appropriate category file (or create a new one) + update `CMakeLists.txt` + this file.
 - If you add a new `.comp` shader → add to `CLPEAK_VK_SHADERS` in `CMakeLists.txt`.
 - If you change `vkPeak` interface → update `include/vulkan/vk_peak.h`.
+- If you change `VulkanDevice` → update `vulkan_device.cpp` + `include/vulkan/vk_peak.h`.
 - If you change `CompileShaders.cmake` → test that `glslc` is found or gracefully skipped.
